requirements.txt
pandas
python-dotenv
google-generativeai
streamlit
openai